requirements.txt
django-filter
asgiref==3.4.1
Django==3.2.6
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
flake8==3.9.2
mccabe==0.6.1
openpyxl==3.0.9
Pillow==8.4.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.3.0
pytz==2021.1
sqlparse==0.4.1
xlwt==1.3.0
